# docs/swagger.yaml
basePath: /
definitions:
  delivery.Delivery:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      status:
        type: string
    type: object
  item.Item:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: number
      rid:
        type: string
      sale:
        type: integer
      size:
        type: string
      status:
        type: integer
      total_price:
        type: number
      track_number:
        type: string
    type: object
  order.Order:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/delivery.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/item.Item'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/payment.Payment'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    type: object
  payment.Payment:
    properties:
      amount:
        type: number
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        type: number
      delivery_cost:
        type: number
      goods_total:
        type: number
      id:
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is simple service for parsing orders
  termsOfService: http://swagger.io/terms/
  title: Orders API
  version: "1.0"
paths:
  /order/{order_uid}:
    get:
      description: get order
      parameters:
      - description: Уникальный номер заказа
        in: path
        name: order_uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный запрос
          schema:
            $ref: '#/definitions/order.Order'
        "400":
          description: Плохой запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Заказ не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: GetOrder
      tags:
      - Order
swagger: "2.0"